requirements.txt
Flask==2.3.3
gunicorn==20.1.0
requests==2.31.0
beautifulsoup4==4.12.2
Pillow==10.0.0
